dev: Move `VerifyFailure` and `FailureLocation` into submodule
